Add files via upload
A2Z is a leading food delivery app in the market. On its 1st anniversary, the management decided to provide a bonus to its employees based on the delivery time. Help the management to find the number of employees who deliver the food on/before the expected time. If none of the employees have delivered the food on time then display all the employee's name along with the message "are not eligible for Bonus".

[The first input is the number of employees who have gone for food delivery, the second series is the name, expected time of delivery (24-hours format), and actual delivered time (24-hours format) by the employees]
Sample Input and Output1:
5

Richard,9:30,9:18

Jessy,18:59,19:01

Sam,9:31,9:30

Williams,9:45,10:10

Annie,20:30,20:59

Richard Sam are eligible for Bonus

Sample Input and Output2:
6

Bella,9:30,9:15

Peter,21:59,21:58

Mohammed,9:10,9:10

Usain,9:20,9:13

Emy,8:30,8:30

Minnie,20:45,20:38

 Bella Peter Mohammed Usain Emy Minnie are eligible for Bonus

Sample Input and Output3:
4

Licy,9:36,9:48

Jeffrey,13:59,14:04

Geoffrey,11:50,12:05

Rosh,22:20,22:31

Licy Jeffrey Geoffrey Rosh are not eligible for Bonus
/////////////////////////////////////////This is the question